Add scenario switching. Use ?scenario=verzikp3 for fun.
