+++ src/sdk/Scenario.js
@@ -1,0 +1,24 @@
+'use strict';
+
+// Base class for any trainer scenario.
+export class Scenario {
+  
+  constructor() {
+  }
+
+  getName() {
+    return "My Scenario";
+  }
+
+  getRegionWidth() {
+    return 29;
+  }
+
+  getRegionHeight() {
+    return 30;
+  }
+
+  // Spawn entities, NPCs, player and initialize any extra UI controls.
+  initialize(region, document) {
+  }
+}
